Remove action page creation from script, rename file
